ON-11912: ON-11912: cp_server: re-ask kernel about new interfaces
When cp_server sees new network interfaces, it should find out their
hwports.  This patch enhances CP_DUMP_HWPORTS ioctl, so that it is
possible to ask about one ifindex only.  This ioctl is used when
cp_server gets a new llap after all hwports have already been dumped.

Also announce hwport when a NIC is re-discovered, which happens with the
new net driver every time the interface goes down-up.

If a NIC was moved to another net namespace and returned back, it is now
properly recognised as acceleratable.

Fixes regression from ON-11912.