ON-11418: When allocating the state shmbuf do flush when unmapping extra pages.
When we allocate the buffer used for the shared state and endpoints we unmap
pages that aren't used for the ci_netif_state. Previously we thought we didn't
need to flush the tlb here, as the pages should not have been used. However,
evidence from testing shows that this flush is necessary.