ON-11919: do not send UDP packet via OS when unbound
When a UDP socket is not bound to any port, Onload used to send the first
packet via OS to get the local port information.  This patch calls bind()
instead and sends the packet in Onload way.

When ECMP (multipath) is used, it means that all the packets are sent via
the same path.