ON-11996: SCTT-1223: Don't log failure to allocate PIO with EPERM.
This is expected on NICs that are not licensed for PIO.  Once upon a time we
used the result of an AppFlex challenge to decide whether to request PIO, but
this is longer an option.  MC_CMD_GET_CAPABILITIES always reports that PIO is
available, even on unlicensed NICs.

This patch suppresses the error message when PIO allocation fails with EPERM,
in NDEBUG builds only.

(cherry picked from commit b531c1bb90926aa27c4472dddd626f3c40187d71)