ON-12031: Reduce dependencies on net driver.
When we import the new virtbus driver the structure of how we import the
net driver will change. In support of that this change reduces the existing
dependencies on the net driver.

This consists of a mixture of removing dependencies that no longer exist,
adding kernel compat to onload for a small number of things where we were
relying on the net driver's compat, and having our own copy of the compat
determining magic functions.