ON-12053: fix race between kill_stack() and oo_netif_apps_gone()
Move n_ep_orphaned back to kernel-only structure in case of non-ulhelper
build.  It allows to ensure that tcp_helper_kill_stack() is able to tell
dtor process "do not care about orphan endpoints", even if kill_stack()
runs before that orphan endpoints are counted by dtor.

Get the orphan-associated refcount when the stack is created.  It serves
2 goals: fixes the race above and removes oo_thr_ref_get() from dtor path.
oo_thr_ref_get() can fail, and it is not easy to handle such failure
on the dtor path.