ON-12085: Allocate AF_XDP rings from efhw
Add a new efhw side interface to be called after user memory has
been provided, to register the memory, create the queues, and make
the AF_XDP socket ready for use. This populates data structures to
allow the queues to be mapped into userspace, and accessed by user
and kernel ef_vi instances.

Remove the vestigial userland calls to mmap, and the AF_XDP file
descriptor that was passed to userland for that purpose.