Cleanup: defer all AF_XDP initialisation until buffer memory is ready
The aim is to reduce the extra work for AF_XDP to a single function
call after buffer memory allocation. However, for the time being we
still need to access the queue memory in order to initialise the VI
before memory is ready, so a vestigial function remains to do that.