ON-12133: Remove onload's use of driverlink hooks by proxying in sfc_resource
When we migrate to virtbus we can only have one client. This change also
removes the concept of driverlink from onload, which only needs to know about
the appearance/disappearance of the net device. This should enable a follow up
change to unify some of the handling of AF_XDP devices.