ON-12250: Don't attempt to map hugetlbfs pages for AF_XDP
Attempting to map individual pages of a hugetlbfs compound page can
corrupt the page table in ways that are difficult to debug. Better
to fail with a simple error before that happens.

Note that this doesn't test for transparent huge pages. I think that
they can probably be used here, though I haven't tried.