ON-12477: Workaround for sketchy AF_XDP redirection in older kernels
Older kernels (pre-5.3) don't support the use of a fallback action,
which we use to pass packets to the kernel if AF_XDP socket lookup
fails. Work around this by looking up the socket before requesting
redirection, if the initial redirection attempt fails.

Unfortunately, older kernels also don't support lookup on an XDP
socket map from a BPF program, so we need a second map of an
acceptable type to shadow the redirection map, and all the
associated resource management shenanigans.