ON-12245: Support drivers requiring explicit location for ethtool filter install
Most NIC drivers do not support automatic selection of a free location for
installation of an ethtool RX filter. This changes uses the ethtool code to
determine a suitable location for the filter.

The ethtool code determines the location by listing all existing rules and then
selecting a free location. There is no concurrency control around this, the
assumption is that ethtool itself is the only thing installing filters. This
means there's the potential for us to select a free location and it being
filled by ethtool if ethtool is also used to install filters at a time when
onload may also be doing so.

The imported ethtool code has been modified to build in our kernel environment.