ON-12585: implement close trampolining via fop->flush and SIGONLOAD
Instead of changing the syscall table (as it was before),
this patch intercepts close() syscalls via flush() operation in
struct file_operation.  For an Onload endpoint we send a SIGONLOAD
signal to the libonload library.

SIGONLOAD is an rt signal, so (1) multiple signals will be queued and (2)
a non-onloaded application ignores this signal by default.

SIGONLOAD is choosen to be from the middle of the RT signal range, so
almost all applications do not use it.  Otherwise the user is free to
redefine the SIGONLOAD value.