ON-12495: Erase erroneous errno extraction
af_xdp_kick() was extracting its return code from errno, but that
is already handled by oo_resource_op(). Doing it twice caused it
to swallow any errors, so we would not know that a further kick
would be needed in future. In some circumstances, that could cause
tx packets to remain queued forever.