ON-12681: rework eplock_lock to take lock in kernel
Really take the stack lock from the locking ioctl, instead of returning
to UL and calling trylock there.

This patch also adds timeout to the stack lock function.  With the new
approach when sleep happens in the same loos as trylock itself it is easy
to calculate the time spent.