ON-12564: kill CI_CFG_CITP_INSIDE_LIB_IS_FLAG flag
We use onload_poll() and other onload function in multiple places, so
CI_CFG_CITP_INSIDE_LIB_IS_FLAG=1 does not work in many places.

And it was completely untested for years.

Also the comment for this option was wrong.  It said that recursive
entering was used for ZC only.  But recursive entering was used in splice
as well.