ON-12538: tune up AF_XDP TX queue kick moderation once more
This is one more try to reduce AF_XDP TX queue kick latency.
See previous commit cf2b2ac17145485aa781fd2beb6c905e702094b8.
On one hand we do not want to kick TX queue too often because of performance
impact. See ON-12129. On the other hand there is a latency problem when waiting
for the TX queue to be stuffed. We should to find out some balance.