ON-12575: Support differing kernel DMA addresses and NIC QDMA addresses
The U26z ARM doesn't use the same DMA addresses that the kernel does. This
patch makes DMAable regions now have two DDR addresses per page: the one that
the kernel gave us (free_addrs, since the only reason we need to hold on to
them is to call dma_unmap_single() at exit) and the ones we should actually use
in descriptors (dma_addrs, as before). The net driver provides a new driverlink
API to do the actual conversion for us, so this patch is essentially just data
management.