ON-13008: Reduce usage of pci_dev.
A lot of places we use pci_dev we could just use the device. At some point we
won't have a pci_dev per interface, so I've started removing the easy cases
now. Some of these are probably only used in cases where we know that we do
have a pci_dev but it's easier to see the other cases if we get rid of as much
as we can.