ON-13031: ef_vi transmit operations for X3/efct
Placeholder for tx-related hardware definitions, which will probably be
replaced by something autogenerated in due course.

Low-level operations to write to the CTPIO aperture, used to partially
implement the "legacy" ef_vi tx operations. If we want a specialised
efct operation, it will look very much like ef_vi_transmitv_ctpio.

This is a work in progress, with various TODOs to be resolved.
In particular, there is no flow control (which will need additional
state).