VIRTBLK-672: Account for plugin DDR base address
Extract the base address of the plugin instance's DDR region from
the reply message and add it to the offset given in the DATA_REMOTE
Ceph message so that upper layers receive the address that they
need in order to issue a mem2mem from the correct DRAM address.

Plugin payload DDR offsets need to wrap around at the end of the
allocated region. This has been implemented in a couple of places
but has not been tested. It's possible that further work is needed.