ON-13092: Make dev_type init handled by resource prober.
How we determine the NIC arch itself depends on the NIC arch, which gives us
a bit of a circular dependency. Logically this might belong as an efhw op, but
we don't know what ops to use until we know the arch. We've dealt with this
before by assuming that anything without a pci_dev is AF_XDP. That's not
sufficient anymore. The resource prober (driverlink, nondl, auxbus) should
know how to init the dev_type for NICs they handle, so shift the dev_type init
there and add it as yet another parameter to efrm_nic_add.