ON-13100: efct support for filtering traffic between kernel and ef_vi/Onload
The net driver will independently check each packet against a filter list
maintained in efhw. There is no need for this filter to match the hardware's
capabilities, but it should be able to express the common use-cases of
customers' code: the net driver and user apps need to be able to agree on who
is going to process each packet, therefore they each need to be able to run
more-or-less the same algorithm. I arbitrarily select EF10's low-latency
firmware's capabilities as that algorithm.

Not currently tested at all.