ON-13198: force stack alignment for vfork_as_fork()
onload___vfork_as_fork() is called from asm code, so we should somehow
guarantee proper stack alignment for SSE operations.  It can be done by
carefully tuning the asm (in vfork_intercept.S), or by forcing alignment
on __vfork_as_fork().

Given that this issue fires in very rare case, it is safer to avoid asm
tuning.  This way we won't meet this issue again if asm is changed for
some unrelated reason.