ON-13310: Make DMA queue memory allocation clearer.
This renames variables to make it clearer where we're dealing with host sized and NIC sized pages. It also removes the allocation of a temporary array to pass the NIC sized pages around. This was wrong anyway in that it sized the array for NIC pages but populated with host pages, but it also introduces an unnecessary failure path. Let's just use the allocated efhw_iopages structure that we've created instead. There was also some redundant and confusing code here, that can just be removed.