ON-13290: ON-13307: dup2: interrupt recv() in other threads
SIGONLOAD is not deferred any more, because it must not be deferred when
it is used for close().  At the same time this signal is used for dup()
to interrupt other threads which are using the same file descriptor, see
bug 28123 / ON-1646.  In such a case it is important to avoid SA_RESTART effect.

This patch ensures that NEED_RESTART flag is removed, and fixes a regression.

Fixes: d02d472e69 ON-13290: do not postpone SIGONLOAD