ON-13326: move proactive resource allocation from unlock to poll
We used to preform proactive socket allocation from slow unlock hook for
the shared stack lock.  Unfortunately slow unlock might be not happening
until it is too late.  This patch moves the proactive check to the end
of the stack poll.

The allocation itself still happens from the unlock hook.