ON-13341: fix a spurious warning about uninitialized data being accessed
ci_dump_stats may access either 32 or 64-bit fields. When the function is
inlined inside stack_dstats, the compiler sees that dstats_t is shorter than
64 bit, but there exists a branch where it is accessed as a 64 bit value, but
it cannot prove the branch is unreachable. For some reason, padding the structure
up to 64 bit exactly is not enough, it need to be larger than that.