VIRTBLK-1262: handle ddr wrap in ceph remote data
When the plugin reaches the end of the DDR ring buffer, the final
message (assuming it doesn't fit exactly at the end of the buffer)
wraps round to the start of the buffer. The plugin still delivers a
single notification to the host in this case.

So we must detect when a message wraps the buffer and deliver two iovs
for the two parts of the message.