ON-13329: Don't fail control plane startup if /var is not present.
In b03eea4, we changed the control plane to create its agent socket in /var/run
rather than in /run for the sake of compatibility with RHEL6.  However, some
systems that do have /run don't have /var early in the boot process.  We no
longer need RHEL6 compatibility, so return to using /run, which is where the
one existing client looks in any case.