ON-13289: respect absence of SA_RESTART in ci_sock_sleep
If ci_sock_sleep() is intercepted by an Onload-postponed signal, it did not
respect SA_RESTART flag and always restarted the blocking syscall.
The issue is extremely hard to reproduce, because the time window for
such an Onload-postponed signal is very short.

Now Onload properly returns -1 with errno=EINTR even if signal comes at
that time.