ON-13023: epoll: keep procmask unchanged during spin+block
Previously Onload exited from the library between spinning and blocking
inside epoll_wait and epoll_pwait.  As a result we changed the sigprocmask
back and forth to keep proper epoll_pwait semantics.  As a side effect,
we failed to return a error with EINTR when a signal comes between spin and
block.

The new approach is to exit from the library inside ioctl.  It means that
all signals are reported properly, and we do not change sigprocmask in the
middle.