ON-13400: Make sure that VIs and IRQs match when rediscovering NICs.
This code dates back to the project in which hotplug support was added for
certain types of fault-tolerant servers.  In those days, it was impossible to
reload the net driver without also reloading Onload, so we were less exposed to
the possibility of the parameters reported by driverlink for an interface
changing across the driverlink-remove and driverlink-add events.

This patch makes the check for compatibility between a new interface and an
unplugged one more stringent.