ON-13400: EF100: Fix crash at stack-creation after net driver reloaded.
On EF100, Onload uses the net driver's mapping of the BAR for poking the
interrupt-prime register.  However, when restoring an interface that has been
removed and then reinserted by driverlink -- as in the case of hotplug or of
net-driver-reload -- we didn't update Onload's cached address within this
mapping.  This patch updates that address, and two other parameters that were
missed similarly.