ON-13111: CTPIO TX compat for efct.
Existing CTPIO using applications expect to have to post a DMA fallback. We
don't need this for efct.

We might want to do something more sophisticated in the future - there was a
suggestion of using the fallback for the the ef_request_id for example - but
for now just keep things simple.