ON-13409: Allow Onload stacks to share interrupt vectors.
On EF100 and EFCT, Onload manages and receives interrupts for its own EVQs.
This is in contrast to EF10, where this is handled by the net driver.  Current
EF100 configurations only allocate a handful of interrupt vectors per PCIe
function, with the result that, unless multiple EVQs can share an interrupt
vector, the number of EVQs that can be allocated is limited severely.

This patch implements that sharing: it introduces an efrm_interrupt_vector
structure into which is factored out the apparatus for managing interrupt
vectors that was previously stored with each efrm_vi, and allows multiple
efrm_vi instances to be associated with a single efrm_interrupt_vector.  This
in turn allows the logic that caps the number of VIs to the number of interrupt
vectors to be removed.