ON-13449: [1/3] New efct superbuf mmap code: old kernels version
Sneakily get hold of hugetlb_file_setup() to get ourselves a
userspace-mmappable hugepage.

The former code required a userspace address space to do its thing, so
got in to difficulty with Onload's work queues: there either wouldn't be
a userspace or we'd use a random userspace and potentially leak
information.