ON-13533: AF_XDP: Fix interrupt-despatch assertion.
The changes for ON-13409 to allow multiple VIs to share an interrupt added an
assert(0) to the error message that was printed when a race was detected at
interrupt-despatch time.  However, such races are in fact legal with AF_XDP,
where waiters are woken by the dinging of the AF_XDP socket's wait queue, which
lacks the serialisation of net-driver wakeups and interrupt tasklets.  This
patch acknowledges that fact.

Fixes: 4613ea6