VIRTBLK-1662: Fix ZC-receive crash on architectures that reorder stores.
When a zero-copy callback indicates to Onload that the application wishes to
retain ownership of the buffer, zc_call_callback() sets the CI_PKT_RX_FLAG_KEEP
bit on the packet, which in effect takes out an extra reference to the packet.
Concurrent stack-pollers will check for this flag when reaping packet buffers
from the receive queue.  As such, it is essential that the KEEP flag be visible
to the stack-poller at the point at which the packet becomes eligible to be
reaped.  This eligiblity-transition occurs for a given packet when the
recv1_extract pointer advances beyond that packet, and therefore the two writes
that must be ordered in this scenario are to the KEEP bit and to recv1_extract.

Those two writes are sequenced correctly in program order, but there was no
intervening write barrier, and block-proxy experiments running on the SN1000's
ARM have fallen foul of this.  This patch adds the necessary barrier.