ON-14021: Introduce new efct rxq stat for superbuf drops
The existing too_many_owned stat can be misleading when the app's
max_allowed_superbufs is less than the net driver's rx ring size, since
that situation means the kernel is constantly bumping up against the app
limit and incrementing the stat. This new stat is incremented at pop
time instead, so should be more representive of when there's a problem.

Not incrementing CI_EFCT_SWRXQ_ABI_VERSION since the new field happens
to fit in to existing padding.