X3-560: Free IRQs if the hardware goes away.
If an efct NIC is removed we need to release all our hardware resources. Most
of this was already covered by our common reset and hotplug code, but we
handle our own IRQs for efct, which wasn't the case for ef10. This change
unbinds the IRQs, but leaves the associated VI infrastructure in place. This is
similar to our handling of other hardware resources that disappear.