VIRTBLK-1920: Fix resetting of new CRC ID after previous insertion failed.
When Onload fails to send a packet with a prefix (most likely because the TXQ
is full), it does some cleanup to revert some state that was changed while
preparing the packet that was not in the end transmitted.  This cleanup had a
bug when reverting packets into which a CRC was to be inserted -- it left the
packet with an associated ID -- with the result that the next CRC-accumulation
to require a new ID was not marked as resetting the CRC of that ID as it ought
to have been.  This patch fixes the cleanup logic.