implement capability to disable netdriver build
Introduce the HAVE_SFC variable to control sfc/driverlink build.
If it is set to 0, disable build of the entire netdriver and
driverlink support, and also other code in sfc_resource module
which uses driverlink API:
- PIO support
- EF100 slice extensions support

It is useful to isolate netdriver and onload modules build to get an easier
way to add support of new kernel versions in case when sfc build is broken
on new kernels (5.18 for now). For AF_XDP we do not need sfc and driverlink,
so we may proceed with new kernel support avoiding netdriver issues.