Fix x86 find_syscall_table() with new IBRS linux code
New enhancements in Spectre mitigation in Linux kernel broke
find_syscall_table() function by adding a few bytes of asm instructions
in the middle of the code we are parsing.

Do not assume that "call do_syscall_64" goes right after saving
the syscall number, to match the code from linux>=5.18.14 and
Debian's backport in 5.10.0-17.