ON-14366: Fix kernel bug attempting to access loopback device with non-X3 driver
Code added to detect use of an unsupported X3 assumes both that the device
pointer is valid, and that the device has a parent. The latter assumption
doesn't hold for a loopback device, and there also seem to be valid code paths
with a null device pointer, so add checks for both cases.