gcc-12: fix segfault when executing library
When the library is run as an executable, new gcc-12 results in segfault.
Simplify the code in onload_version_msg(), so that it works seamlessly.

Avoid local_strlen(), because it results in strlen() call in any way
(when using gcc-12).
Additionally, msg0 was not initialized properly with gcc-12, and I failed
to understand why.  Given that local_strlen() is harmful now, I did not
explore it further.