ON-14420: remove assertion from ef_vi_stats_query_layout
We shouldn't fail an assertion if this is called for an unsupported
architecture, just return an error code. (EINVAL doesn't really make sense,
but we have existing code that interprets this value to mean "unsupported
architecture" so I'm not going to try to change it).

The similar assertion in ef_vi_stats_query should stay, as it is an error
to call that function if the layout is unknown.