ON-14312: Skip test_singleroute_transparent() by rearranging test calls
Make sure cpdecorate() applies to the "internal" test functions prefixed
with "do_" to (1) achieve consistency with other tests and (2) workaround
its wrapping behaviour that doesn't preserve the wrapped function's
properties, e.g. with functools.wraps(). Consequently, it makes pytest
skip the failing test_singleroute_transparent() test as expected.

Alternatively, we could fix cpdecorate() itself, but we would also need
to parametrise the arguments it passes to the wrapped functions, which
would be a more involved change.