ON-14503: Disable SFC Siena support
This patch disables the Siena support by unsetting CONFIG_SFC_SIENA
with the brand new ONLOAD_MAKEFLAGS in the top-level Makefile and
passthruparams in the mmake build tree.

Consequently, it addresses a build time error against the Linux kernel
5.19+, which introduces a new option with the same name and meaning and
causes the Onload/SFC to take an unsupported build path if Siena support
is enabled in the config.