ON-14121: fix onload_iptables interface encoding issue
Previously these were read by Popen as bytes. This lead to the
interfaces being converted to strings as e.g. b'p2p1' instead of just
p2p1. This meant when the rules are added by onload_iptables, they get
added incorrectly, resulting in messages like the following:
> efrm_text_to_table_entry: Adding rule for unknown interface b'p2p1'.
which breaks everything as the interface names are now completely
incorrect.
Fix this by setting the Popen encoding to ascii, to decode to a string
correctly.