ON-14527: Fix AF_XDP VI constraints predicate
The IRQ channel may or may not be specified with the EF_IRQ_CHANNEL
environment variable. If it is, we accept only the selected channel.
Otherwise, accept any. Emit a warning if unsupported features are
requested.

Before the fix, the predicate refused any channel if EF_IRQ_CHANNEL was
unset (or when avc->channel was -1), consequently refusing to start
AF_XDP onloaded apps.

Thanks-to: Siân James <sian.james@xilinx.com>