ON-14528: use printf instead of ci_log_nonl for large buffered output
ci_log_nonl has a maximum output of 511 chars (due to CI_LOG_MAX_LINE
limit in ci_vlog). This makes it fine for all of the cases where we are
using it to interleave small bits of output with other ci_log calls.
However, in dump_via_buffers, we simply print a large buffer of
pre-formatted output to stdout. This often exceeds 511 characters, and
truncation occurs. As this is the only output here, printf shouldn't
cause any ordering issues due to lack of interleaving with ci_log calls.