ON-14550: Move pio_buddy_test tools/ip -> tests/onload/buddy
Move the buddy allocator unit tests to the tests directory where it
should have been placed in the first place. It also addresses a build
time error for the tarballs created with onload_mkdist:

  $ mmakebuildtree --user
  $ make -C "$(mmaketool --toppath)/build/$(mmaketool --userbuild)"
  $ make[2]: *** No rule to make target 'pio_buddy_test.o', needed by 'pio_buddy_test'.  Stop.

These tarballs don't contain the required file because it's an internal
unit test. When we move it into a new directory, we start relying on the
Make[All]Subdirs that tolerates the absent directories and thus resolves
the initial error for the cases when the unit test file doesn't exist.