udp_recv: Return real length if MSG_TRUNC is set in flags
In man recv(2):
  MSG_TRUNC (since Linux 2.2)
    For raw (AF_PACKET), Internet datagram [...] sockets: return
    the real length of the packet or datagram, even when it was
    longer than the passed buffer.

This behavior is missing in onload prior to this patch, and onload
would rather return the truncated size of the buffer, if the packet
is larger than the provided buffer.

Code is wrapped in HAVE_MSG_FLAGS because flags cannot be non-zero
in kernel mode.

Note that TCP is unaffected by this and has a different behavior for
MSG_TRUNC.