ON-14637: Don't share PDs for AF_XDP clusters.
This retains the pd for the cluster, but no longer allocates buffer memory for
it, instead allocating new pds for each vi in the cluster. This is necessary
on AF_XDP where the umem memory allocation is specific to a given queue.

Alternatively we could probably avoid creating a pd at all here, but that would
make the AF_XDP case exceptional. We care less about wasting a pd than we do
about allocating a bunch of extra buffer memory, which is particularly bad in
the AF_XDP case as we always allocate the maximum allocation up front.