ON-14813: Extend page offset type in hugetlb library
The patch changes the *byte* offset field type in the huge page
allocator from 32-bit off_t to 64-bit loff_t.

The 32-bit (signed) offset effectively means 2GiB of memory or 1024x
2MiB huge pages. This limit is easy to achieve:

  $ echo 2048 | sudo tee /proc/sys/vm/nr_hugepages
  $ sudo modprobe onload max_packets_per_stack=2008000
  $ EF_MAX_PACKETS=2000000 onload nc -l -p 1234
  $ onload_stackdump alloc_pkts 2000000

The above alloc_pkt command allocates ~2000x hugepages and overflows
the 32-bit offset, causing mapping issues in the kernel space and
leaving it in a semi-broken state.

To address this, the page converts the byte offset to 64-bit type.