ON-14816: Fix hugepage release for atomic contexts
This patch adds the new parameter atomic_context to the
oo_hugetlb_page_free*() functions. It controls whether it is safe to
perform non-atomic operations, in particular, call vfs_fallocate(), that
can sleep.

Currently, there are two users of the hugetlb library: AUX client/EFCT
and iobufset as part of tcp_helper_resource. The former almost certainly
releases huge pages in the softirq context initiated by call_rcu(), so
we disable non-atomic calls here. The latter likely allows non-atomic
calls, but we still play safe and dynamically check if we are in an
atomic context.

The patch identifies that the code runs in the atomic contexts with the
in_atomic() and in_interrupt() predicates that are not recommended.
Instead, we should consider in_task() or others from <linux/preempt.h>.
However, since in_atomic() and in_interrupt() are widely used in Onload
and iobufset in particular, I prefer them until we replace all these in
one go.