ON-14799: EFCT transmit warming
Implement warming of the EFCT transmit path, which is identical to a regular
send apart from setting a flag in the transmit header.

Provide a semi-public EFCT-specific API to use it. No attempt to provide
compatibility with the existing (internal) interface at this time.

Requires a corresponding change to tcpdirect to use this interface for EFCT.