syscall_x86: Allow non-RAX indirect jumps
In a clang-compiled Linux we are testing with, the syscall retpoline
jump was compiled to:

  4c 8b 1c cd 90 02 20 82     mov    -0x7ddffd70(,%rcx,8),%r11
  4c 89 f7                    mov    %r14,%rdi
  e8 ad 5d 3b 00              call   <x86_indirect_thunk_r11>

Because R11 is an x86_64 extension register, emit_indirect()
gives the modified call instruction an REX.B prefix (0x41), becoming:

  41 ff d3                    call   *%r11
  66 90                       nop2

This patch adds a condition for such. It also allows other
non-extension-registers that isn't RAX to be used, instead of
hardcoding 0xd0 for RAX.

I believe in the case of X86_FEATURE_RETPOLINE_LFENCE with extension
registers, the patch simply won't be performed, because lfence + call
now takes 6 bytes and apply_retpolines skips the patch when the
length does not match. I have not tested this this case yet however.