ON-15222: Initial efx_auxiliary resource driver support
This puts some scaffolding in place suffficient to allow probe and remove of
an ef10ct device. As it stands the code has various ifdefs and contains bits
that need refactoring to reflect commonality with other arches. The intent at
this point is to get something into a branch where others can work on it.