af_xdp: support external XDP programs
Onload currently manually installs its XDP map and program, in
__af_xdp_nic_init_hardware.

Support more advanced programs and use standard XDP tools.

Modify __af_xdp_nic_init_hardware to check whether a (pinned) map with
the expected name already exists. If so, take a reference to that map
instead of creating a new map.

The code expects that if this map exists, the matching XDP program
also exists and has been attached to the device.

Signed-off-by: Willem de Bruijn <willemb@google.com>