ON-14825: poll interfaces cyclically
Interface polling was done in the same order every time which led to
some interfaces getting preferential treatment over others, notably
in the difference between the time a packet was received at the NIC
and the time it was serviced by the application. This change greatly
reduces the systematic favouritism by cyclically selecting the first
interface to poll then linearly polling the remaining interfaces.