driver/linux_onload: remove wrong sendpage assertion
Onload used to assert that a memory chunk passed to sendpage handler
is always within one page.  It was true at the time when the assertion
was added.  It is not true any more.

If a user runs two splice() calls: OS socket -> OS pipe -> Onload
socket, then Onload's sendpage handler gets a whole memory-continuous
packet as it was received by the NIC.  Such a packet tend to be a part
of a compound page, and it may cross the bounds of an ordinary page.

The real sense of this assertion was "ensure that we are handling
a continuous chunk".  This is still true.  But the assertion is bad for
linux>=6.1 and for some updates of older kernels.

OL-Redmine-Id: 12010
Co-developed-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>
Signed-off-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>
Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>