lib/efthrm: fix packet injection to kernel
Ensure that a list of packets to be injected to kernel is properly
detached from the stack structures, so that it is not updated under
the feet of the deferred work.

Also add a guard to avoid dead loop in kernel when shared state
is corrupted.

OL-Redmine-Id: 13238
Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Sergey Nikitin <sergey.nikitin@oktetlabs.ru>
(cherry picked from commit ae21e12bd23b9c09bf2ff02e2861e1529d19fd7b)