ON-12878: Expose bond master ifindex through the MIB
It would have been possible to expose the entire bond table through the
MIB instead, i.e. revert 7439705ca1, but that includes a whole bunch of
irrelevant stuff and brings MIB_UPDATE_LOOP complexity to that data
structure, which is annoying. Just adding the field we need to the
existing LLAP table is easier.