driver: fix out-of-bounds UBSAN warning
Since commit 2d47c6956ab3 ("ubsan: Tighten UBSAN_BOUNDS on GCC")
in Linux-6.5 we can't use arrays of size [0] or [1] because it
triggers UBSAN false-positive warnings "array-index-out-of-bounds".

Use __DECLARE_FLEX_ARRAY() macro to declare flexible arrays in those
cases. For arrays with size [1] use a union to keep the size of a
struct containing the flexible array unchanged.

Signed-off-by: Sergey Nikitin <sergey.nikitin@oktet.tech>