ON-15328: Add lock to oo_hugetlb_allocator to serialise allocations
The initial implementation of the hugetlb kernel util library did not
implement any serialisation for hugepage allocations. This could
potentially lead to memory corruption and crashes if multiple threads
tried to allocate at the same time.

This commit adds a lock to the oo_hugetlb_allocator struct which is used
when allocating new hugepages, and so should prevent the issues seen
previously.