ON-15410: fix invalid nanoseconds in tx timestamps
Overlaying the new "fractional bits" field with the old "format" field
resulted in it being overwritten by initialisation code which expects
to set the old field, causing an incorrect shift value and bogus results.

The simplest fix is to make the new field separate. Hopefully this doesn't
change the layout of "struct ef_vi" in a way that impacts performance.