ON-15252: Avoid shm access in rollover check
We need to decouple EFCT polling from the X3-specific shared memory
structure. For the rollover check, this can be done by storing an
"end" value in the "rxq_ptr" structure and comparing with that,
avoiding shm access altogether. This may also be a slight optimisation
as it has better locality and avoids a bit of arithmetic.