ON-15395: Add multiarch NIC support to cplane
This patch paves the way to support the multiarch NICs, those that have
two hwports per interface.

It breaks the corresponding assumption in the code: that a non-bonded
"plain" interface can have only one associated hwport, but surprisingly,
it does not require too many changes to accommodate it.

To ensure the correctness, it extends the system- and cplane unit tests
to include multiarch NICs.