linux_onload: do not lock RCU for alloc_fd
There is no need to lock RCU before get_unused_fd_flags() (which boils
down to alloc_fd()).  The comment sais that RCU lock in needed for
files_fdtable(), but it is not true.  alloc_fd() locks the file_lock
spinlock, so it is OK without any additional RCU.

On the flip side, alloc_fd() may call expand_files() which may sleep.
So this RCU is both unnecessary and harmful.

Surprisingly it is the only call for get_unused_fd_flags() with RCU lock
decoration.  All the other calls are already good, without any
additional RCU call.

OL-Redmine-Id: 13489
Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>