ON-15395: Add test coverage for cp_get_hwport_ifindex
cp_get_hwport_ifindex is quite an untrivial function used in TCPDirect,
i.e. part of the interfacing with other software. However, it lacks any
test coverage.

This patch extends the test_namespace cplane unit tests to call
cp_get_hwport_ifindex from within a secondary namespace to check that it
does not cross the namespace boundary and returns the best matching
interface, which is not the originating plain/bare-metal one from the
main namespace.

Curiously, it is supposed to return CI_IFID_BAD for veth in the
secondary namespace but would return a genuine veth interface number if
there is only one registered hwport in the main namespace.

Additionally, it extends the cplane system tests to call
cp_get_hwport_ifindex immediately after creating a bond.