ON-14055: Fix stack creation after NIC unplug
This is necessary but not sufficient to fix hotplug. When a NIC is unplugged
we retain state within bob. This should be used to create ghost VIs if stacks
are created while the device is absent, which can later be un-ghosted if the
NIC reappears. However, the dmaxlate path requires use of driverlink, which we
can't do for a device that's not there.

This change unbreaks this particular problem introduced with the virtblk
support and allows us to create stacks again after a NIC has been unplugged,
but there remain other issues preventing full hotplug support from working at
the moment.