driver/onload: simplify compat code for oo_fd_replace_file()
Replace linux-version-dependent code for obtaining file pointer
out of the descriptor with simple fget(). It allows not to deal
with different linux versions.

fget() takes a file reference, so we must do extra fput().
Since the new `lookup_fdget_rcu()`, which we were supposed
to use, do it either, there is no reason to use it.
Call fget() instead and remove all the related compat code.

Fixes: e34a911603af ("driver: compat for lookup_fd_rcu()")
Signed-off-by: Sergey Nikitin <sergey.nikitin@oktet.tech>